chore: refactor pool provider into postgres module
